+++ pkg/agent/as3/as3Manager.go
@@ -36,9 +36,9 @@ const (
 	svcPoolLabel        = "cis.f5.com/as3-pool="
 	as3SupportedVersion = 3.18
 	// Update as3Version,defaultAS3Version,defaultAS3Build while updating AS3 validation schema.
-	// While upgrading version update $id value in schema json to https://raw.githubusercontent.com/F5Networks/f5-appsvcs-extension/master/schema/latest/as3-schema.json
-	as3Version           = 3.50
-	defaultAS3Version    = "3.50.0"
+	// While upgrading version update $id value in schema json to https://raw.githubusercontent.com/F5Networks/f5-appsvcs-extension/main/schema/latest/as3-schema.json
+	as3Version           = 3.52
+	defaultAS3Version    = "3.52.0"
 	defaultAS3Build      = "5"
 	as3tenant            = "Tenant"
 	as3class             = "class"
@@ -46,13 +46,13 @@ const (
 	as3application       = "Application"
 	as3shared            = "shared"
 	as3template          = "template"
-	// as3SchemaLatestURL   = "https://raw.githubusercontent.com/F5Networks/f5-appsvcs-extension/master/schema/latest/as3-schema.json"
+	// as3SchemaLatestURL   = "https://raw.githubusercontent.com/F5Networks/f5-appsvcs-extension/main/schema/latest/as3-schema.json"
 	as3defaultRouteDomain = "defaultRouteDomain"
-	as3SchemaFileName     = "as3-schema-3.50.0-5-cis.json"
+	as3SchemaFileName     = "as3-schema-3.52.0-5-cis.json"
 )
 
 var baseAS3Config = `{
-	"$schema": "https://raw.githubusercontent.com/F5Networks/f5-appsvcs-extension/master/schema/%s/as3-schema-%s.json",
+	"$schema": "https://raw.githubusercontent.com/F5Networks/f5-appsvcs-extension/main/schema/%s/as3-schema-%s.json",
 	"class": "AS3",
 	"declaration": {
 	  "class": "ADC",