+++ docs/RELEASE-NOTES.rst
@@ -10,8 +10,9 @@ Added Functionality
 **What's new:**
     * Multi Cluster
     * CRD
-        * `Issue 3471 <https://github.com/F5Networks/k8s-bigip-ctlr/issues/3471>`_ Support for loadBalancerClass for service type lb. See `Example <https://github.com/F5Networks/k8s-bigip-ctlr/blob/2.x-master/docs/config_examples/customResource/serviceTypeLB/>`_
-    * `Issue 3430 <https://github.com/F5Networks/k8s-bigip-ctlr/issues/3430>`_ Support for CIS deployment parameters "ipam-namespace" to configure the namespace for IPAM CR
+        * `Issue 3471 <https://github.com/F5Networks/k8s-bigip-ctlr/issues/3471>`_: Support for loadBalancerClass for service type lb. See `Example <https://github.com/F5Networks/k8s-bigip-ctlr/blob/2.x-master/docs/config_examples/customResource/serviceTypeLB/>`_
+        * `Issue 3438 <https://github.com/F5Networks/k8s-bigip-ctlr/issues/3438>`_: Support for FTP Profile in Policy CR. See `Example <config_examples/customResource/Policy/policy-with-ftp-profile.yaml>`_
+    * `Issue 3430 <https://github.com/F5Networks/k8s-bigip-ctlr/issues/3430>`_: Support for CIS deployment parameters "ipam-namespace" to configure the namespace for IPAM CR
 Bug Fixes
 ````````````
 * `Issue 3401 <https://github.com/F5Networks/k8s-bigip-ctlr/issues/3401>`_: Fix for invalid iRule generation for HTTP/2 full proxy mode